Update README.md
From Epidemic Statistic to Boardroom Executive Blueprint. 

And we are only getting started. 

This is what happens when you self-realize thru spiritual enlightenment after having unjustly been sent on wilderness walks for the better part of seven years. 
I implore you to stay tuned. Or do as the rest, and dismiss. 

#YHWH 
#ExecutiveSearch
#AllWays